flask-server/requirements.txt
Flask
flask-cors
flask-mysqldb
bcrypt
python-dotenv
PyJWT
